Fix printf type safety in RGB panel logs
Add explicit casts for `DEBUG_PRINTF`/`printf` arguments in RGB panel code to match the expected format specifiers (`%u`, `%x`, `%p`). This avoids format-type mismatches and compiler warnings while keeping runtime behavior unchanged.

+++ lib-rgbpanel/src/rgbpanel.cpp
@@ -216,5 +216,5 @@ void RgbPanel::Stop() {
 
 void RgbPanel::Print() {
     puts("RGB led panel");
-    printf(" %ux%ux%u\n", columns_, rows_, chain_);
+    printf(" %ux%ux%u\n", static_cast<unsigned>(columns_), static_cast<unsigned>(rows_), static_cast<unsigned>(chain_));
 }
